+++ packages/common/src/index.ts
@@ -9,3 +9,9 @@ export * from './data-converter-helpers';
 export * from './patch-protobuf-root';
 export * from './tls-config';
 export * from './utils';
+export {
+  ProtobufBinaryPayloadConverter,
+  ProtobufJsonPayloadConverter,
+  DefaultPayloadConverterWithProtobufs,
+  DefaultPayloadConverterWithProtobufsOptions,
+} from '@temporalio/workflow-common/lib/converter/protobufs';
