+++ packages/test/src/payload-converters/payload-converter.ts
@@ -1,7 +1,7 @@
-import { DefaultPayloadConverter } from '@temporalio/common';
+import { DefaultPayloadConverterWithProtobufs } from '@temporalio/common';
 import root, { foo } from '../../protos/root';
 
 // Used in tests
 export const messageInstance = foo.bar.ProtoActivityInput.create({ name: 'Proto', age: 1 });
 
-export const payloadConverter = new DefaultPayloadConverter({ protobufRoot: root });
+export const payloadConverter = new DefaultPayloadConverterWithProtobufs({ protobufRoot: root });
